fix[ENTESB-13149]: Add test results to custom resource
YAKS custom resource that represents a test instance is now enriched with test results. The test results are read from the termination log and get parsed as Json blob into the custom resource. Users can review the test results on the custom resource or on the pod via oc client then.

- Always use jackson library for Json marshal/unmarshal operations in all modules
- Add additional printer columns for test results on test CRD
- Change termination log to hold the test results as Json object
- Add test results to test types and fill with results from termination log