#118 Allow yaks-viewer role to manage Knative resources
Adjust yaks roles to be able to create, update, delete Knative resources such as triggers, brokers, channels and subscriptions. This way the event consumption is made possible where
a trigger or subscription is created to point to the Http service that is running as part of the YAKS test Pod.