fix(#159): Reset test status on update
Introduce new test phase updating so operator is not handling test changes anymore. In this phase update the test status and the test spec when updating an existing test. This makes sure to reset the test status properly on the update. Also might fix #74 where status updates and operator changes interfere with each other.